Renamed 'FakeHome' files to 'About', updated route
