+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import Home from './routes/Home'
-import FakeHome from './routes/FakeHome';
+import About from './routes/About';
 import HeaderBar from './components/HeaderBar'
 import './index.css'
 import {
@@ -21,7 +21,7 @@ const router = createBrowserRouter([
   },
   {
     path: "/about",
-    element: <FakeHome />,
+    element: <About />,
   },
   {
     path: "/contact",
